move tests in iblphotometry_tests package
