+++ src/iblphotometry/synthetic.py
@@ -7,7 +7,6 @@ import pandas as pd
 import scipy.signal
 
 
-
 def synthetic101(fs=30, rl=1000, event_rate=0.2):
     """
     Generates synthetic photometry data with
@@ -47,8 +46,12 @@ def generate_dataframe(sigma: float = 0.01):
         df[['raw_calcium', 'raw_isosbestic']] += np.random.randn(*df.shape) * sigma
 
     raw_dfs = dict(
-        raw_calcium=pd.DataFrame(df['raw_calcium'].values, index=df.index, columns=['Region01']),
-        raw_isosbestic=pd.DataFrame(df['raw_isosbestic'].values, index=df.index, columns=['Region01']),
+        raw_calcium=pd.DataFrame(
+            df['raw_calcium'].values, index=df.index, columns=['Region01']
+        ),
+        raw_isosbestic=pd.DataFrame(
+            df['raw_isosbestic'].values, index=df.index, columns=['Region01']
+        ),
     )
 
     return raw_dfs
