+++ src/iblphotometry_tests/data/77a6741c-81cc-475f-9454-a9b997be02a4/Region3G/raw_photometry.pqt
