+++ src/iblphotometry_tests/data/raw_photometry.csv
