+++ src/iblphotometry_tests/data/wittenlab/Subjects/fip_40/2023-05-18/001/alf/photometry/photometryROI.locations.pqt
