+++ src/iblphotometry_tests/test_plots.py
